add self-defined class MyFAISS
