+++ vectorstores/__init__.py
@@ -1,0 +1,1 @@
+from .MyFAISS import MyFAISS
